% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{available_models}
\alias{available_models}
\title{Available models for prediction}
\usage{
available_models()
}
\description{
Available models for prediction using \url{decrypt()}.
}
\details{
Currently, available captchas are:

\enumerate{
\item \href{https://pje.trf5.jus.br/pje/ConsultaPublica/listView.seam}{trf5}: Tribunal Regional Federal 5
\item \href{https://www4.tjmg.jus.br/juridico/sf/proc_resultado.jsp?comrCodigo=24&numero=1&listaProcessos=50718889720218130024&btn_pesquisar=Pesquisar}{tjmg}: Tribunal de Justiça de Minas Gerais
\item \href{https://pje-consulta.trt3.jus.br/pje-consulta-api/api/processos/2104879}{trt}: Tribunal Regional do Trabalho 3
\item \href{http://esaj.tjba.jus.br/cpopg/open.do}{esaj}: Tribunal de Justiça da Bahia
\item \href{https://www.jucesponline.sp.gov.br/ResultadoBusca.aspx}{jucesp}: Junta Comercial de São Paulo
\item \href{https://srv01.tjpe.jus.br/consultaprocessualunificada/}{tjpe}: Tribunal de Justiça de Pernambuco
\item \href{https://www.tjrs.jus.br/site_php/consulta/verificador.php}{tjrs}: Tribunal de Justiça do Rio Grande do Sul
\item \href{https://www.cadesp.fazenda.sp.gov.br/(S(vyfz1cfybbxj3sgpf4eqhxd3}{cadesp}: Centro de Apoio ao Desenvolvimento da Saúde Pública
\item \href{https://sei.economia.gov.br/sei/modulos/pesquisa/md_pesq_processo_pesquisar.php?acao_externa=protocolo_pesquisar&acao_origem_externa=protocolo_pesquisar&id_orgao_acesso_externo=0}{sei}: Sistema Eletrônico de Informações - ME
\item \href{https://servicos.receita.fazenda.gov.br/servicos/cnpjreva/Cnpjreva_Solicitacao_CS.asp}{rfb}: Receita Federal
}
}
