% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{captcha_ds_in_memory}
\alias{captcha_ds_in_memory}
\title{Captcha in-memory dataset}
\usage{
captcha_ds_in_memory(files, dims, output_dim, transform = identity)
}
\arguments{
\item{files}{captcha files}

\item{dims}{vector of length two with file dimentions}

\item{output_dim}{output (y) dimensions}

\item{transform}{transform function}
}
\description{
Captcha in-memory dataset
}
