% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{captcha_fit_model}
\alias{captcha_fit_model}
\title{Fit Captcha model}
\usage{
captcha_fit_model(
  dir,
  dir_valid = NULL,
  prop_valid = 0.2,
  dropout = 0.25,
  dense_units = 200,
  decay = 0.99,
  batch_size = 40,
  epochs = 100
)
}
\arguments{
\item{dir}{directory where the classified images are}

\item{dir_valid}{(optional) directory to validation files}

\item{prop_valid}{proportion of total images considered to validation. Default 0.2.}

\item{dropout}{dropout hyperparameter. Default 0.25.}

\item{dense_units}{number of dense units to use after convolution steps. Default 200.}

\item{decay}{Weight decay applied each epoch.}

\item{batch_size}{Minibatch size. Default 40.}

\item{epochs}{Number of epochs to use. Default 100. The model uses early
stopping, so it is possible that the procedure ends before the total
number of epochs actually run.}
}
\value{
fitted model of class \code{luz_module_fitted}
}
\description{
Fit Captcha model
}
