% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{captcha_load_model}
\alias{captcha_load_model}
\title{Load captcha model}
\usage{
captcha_load_model(captcha)
}
\arguments{
\item{captcha}{file name or captcha name from \url{available_models()}}
}
\description{
Load captcha model
}
