% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{captcha_transform_image}
\alias{captcha_transform_image}
\title{File to torch tensor}
\usage{
captcha_transform_image(x, input_dim = c(32L, 192L))
}
\arguments{
\item{x}{file path}

\item{input_dim}{resize image to dimension}
}
\description{
File to torch tensor
}
