% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{captcha_transform_label}
\alias{captcha_transform_label}
\title{File to response matrix (tensor)}
\usage{
captcha_transform_label(all_letters, vocab)
}
\arguments{
\item{all_letters}{list of tokens for all files}

\item{vocab}{unique tokens}
}
\description{
File to response matrix (tensor)
}
