% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{fit}
\alias{fit}
\alias{train_step}
\alias{valid_step}
\alias{update_train_metrics}
\alias{update_valid_metrics}
\alias{print_results}
\title{Model fitting}
\usage{
train_step(b, model, optimizer, device = "cuda")

valid_step(b, model, device = "cuda")

update_train_metrics(metrics, res)

update_valid_metrics(metrics, res)

print_results(epoch, train_metrics, valid_metrics)
}
\arguments{
\item{b}{batch}

\item{model}{model}

\item{optimizer}{optimizer}

\item{device}{"cuda" or "cpu}

\item{metrics}{metrics}

\item{res}{results}

\item{epoch}{epoch to print}

\item{train_metrics}{train metrics to print}

\item{valid_metrics}{valid metrics to print}
}
\description{
Models related to model fitting
}
