% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{models}
\alias{models}
\alias{captcha_available_models}
\alias{captcha_load_model}
\title{Load or list available models for prediction}
\usage{
captcha_available_models(repo = "decryptr/captcha", tag = "captcha_model")

captcha_load_model(captcha, repo = "decryptr/captcha", tag = "captcha_model")
}
\arguments{
\item{repo}{repository in the form \code{"<user>/<captcha>"}.
Defaults to \code{"decryptr/captcha"}}

\item{tag}{tag name of the release to load the file.}

\item{captcha}{file name or captcha name}
}
\value{
\code{captcha_available_models()} returns a character vector with the
names of the available models. \code{captcha_load_model()} returns an object
of class \code{luz_module_fitted}.
}
\description{
The \code{captcha_available_models()} function lists the models available for
prediction in the \code{{captcha}} package or another repository referenced
in the \code{repo} parameter.
The models are loaded from the \href{https://github.com/decryptr/captcha/releases/tag/captcha_model}{captcha repository release page}
using the \code{\link[piggyback:pb_list]{piggyback::pb_list()}} function.
The \code{captcha_load_model()} function downloads a model using
\code{\link[piggyback:pb_download]{piggyback::pb_download()}} and returns a \code{luz_module_fitted} object.
}
\details{
Currently, available models are:
\itemize{
\item \href{https://pje.trf5.jus.br/pje/ConsultaPublica/listView.seam}{trf5}:
Tribunal Regional Federal 5
\item \href{https://www4.tjmg.jus.br/juridico/sf/captcha.svl}{tjmg}:
Tribunal de Justiça de Minas Gerais
\item \href{https://pje-consulta.trt3.jus.br/consultaprocessual/}{trt}:
Tribunal Regional do Trabalho 3
\item \href{http://esaj.tjba.jus.br/cpopg/open.do}{esaj}:
Tribunal de Justiça da Bahia
\item \href{https://www.jucesponline.sp.gov.br/ResultadoBusca.aspx}{jucesp}:
Junta Comercial de São Paulo
\item \href{https://srv01.tjpe.jus.br/consultaprocessualunificada/}{tjpe}:
Tribunal de Justiça de Pernambuco
\item \href{https://www.tjrs.jus.br/site_php/consulta/verificador.php}{tjrs}:
Tribunal de Justiça do Rio Grande do Sul
\item \href{https://sei.economia.gov.br/infra_js/infra_gerar_captcha.php?codetorandom=112-53}{sei}:
Sistema Eletrônico de Informações - ME
\item \href{https://solucoes.receita.fazenda.gov.br/Servicos/cnpjreva/captcha/gerarCaptcha.asp}{rfb}:
Receita Federal do Brasil
}

This list may be updated when new models are added to the release.
}
\examples{
# run only with internet connection
if (interactive() && torch::torch_is_installed()) {

  captcha_models_available()

  # loads rfb model
  model <- captcha_load_model("rfb")
  # model estimated accuracy on validation dataset
  utils::tail(model$records$metrics$valid, 1)[[1]][["captcha acc"]]

}


}
