% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.R
\name{new_captcha}
\alias{new_captcha}
\title{Create a new project to solve a custom captcha}
\usage{
new_captcha(path)
}
\arguments{
\item{path}{A path. If it exists, it is used. If it does not exist,
it is created, provided that the parent path exists.}
}
\description{
Create a new project to solve a custom captcha
}
