% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{prepare}
\alias{prepare}
\alias{train_transforms}
\alias{valid_transforms}
\alias{captcha_ds_in_memory}
\title{Prepare image}
\usage{
train_transforms(img)

valid_transforms(img)

captcha_ds_in_memory(files, dims, output_dim, transform = identity)
}
\arguments{
\item{img}{image}

\item{files}{captcha files}

\item{dims}{vector of length two with file dimentions}

\item{output_dim}{output (y) dimensions}

\item{transform}{transform function}
}
\description{
Prepare image

Prepare image
}
