% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{captcha_annotate}
\alias{captcha_annotate}
\title{Annotate captchas with their labels}
\usage{
captcha_annotate(files, labels = NULL, path = NULL, rm_old = FALSE)
}
\arguments{
\item{files}{Either an object of class \code{captcha} or a character vector
with the paths to captcha files}

\item{labels}{Either \code{NULL} (for interactive classification) or
a character vector with labels for the Captchas}

\item{path}{Where to save the annotated captcha files.
If \code{NULL}, saves the files in the same folder the unanswered counterparts.}

\item{rm_old}{Whether or not to delete unanswered captchas after
copying and renaming them}
}
\value{
A character vector with the paths to the newly created files
}
\description{
Given one or more Captchas, this function
prompts you to solve them mannually so that later you can train
a model with those labels. Annotated captchas are saved at \code{path}
with their labels in the filename separated by an underscore.
}
