% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{captcha_dataset}
\alias{captcha_dataset}
\title{Captcha datasets}
\usage{
captcha_dataset(
  root,
  transform_image = captcha::captcha_transform_image,
  transform_label = captcha::captcha_transform_label,
  augmentation = NULL
)
}
\arguments{
\item{root}{(string): root directory of dataset where \code{captcha.zip}
exists or will be saved to if download is set to \code{TRUE}}

\item{transform_image}{(callable, optional): A function/transform
that takes in an file path and returns an torch tensor prepared
to feed the model.}

\item{transform_label}{(callable, optional): A function/transform
that takes in the file paths and transform them.}

\item{augmentation}{(function, optional) If not \code{NULL}, applies a
function to augment data with randomized preprocessing layers.}
}
\description{
Captcha datasets
}
