% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{captcha_load_model}
\alias{captcha_load_model}
\title{Load captcha model}
\usage{
captcha_load_model(captcha, repo = "decryptr/captcha", tag = "captcha_model")
}
\arguments{
\item{captcha}{file name or captcha name}

\item{repo}{repo in the form \code{"<user>/<captcha>"}. Default \code{"decryptr/captcha"}}

\item{tag}{tag name of the release to load the file.}
}
\description{
Load captcha model
}
