% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{captcha_transform_image}
\alias{captcha_transform_image}
\title{File to torch tensor}
\usage{
captcha_transform_image(x, input_dim = c(32L, 192L))
}
\arguments{
\item{x}{character vector with the paths to image files.}

\item{input_dim}{resize image to dimension. Defaults to 32x192, which is
a good default for many Captcha applications.}
}
\value{
torch tensor with dimensions \code{length(x)}x\code{3}x\code{input_dim}.
}
\description{
This function uses the \code{torchvision} package to read and transform the
image in a torch tensor. The function tries to adjust the dimensions to
deal with black and white or coloured images.
}
\examples{

if (torch::torch_is_installed()) {
  captcha_file <- fs::dir_ls(
    system.file("examples/captcha/", package = "captcha"
  ))
  result <- captcha_transform_image(captcha_file)
  class(result)
  dim(result)
}

}
