% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{captcha_transform_label}
\alias{captcha_transform_label}
\title{File to response matrix (tensor)}
\usage{
captcha_transform_label(all_letters, vocab)
}
\arguments{
\item{all_letters}{list of tokens for all files}

\item{vocab}{unique tokens}
}
\value{
torch tensor with dimensions \code{length(all_letters)}x\code{length(vocab)}
containing only zeros and ones. All rows sum exactly one.
}
\description{
This function performs a one-hot encoding of the label, transform a label
with \code{N} letters in a matrix of dimensions \code{N}x\code{length(vocab)}. All the
labels must have the same length.
}
\examples{
vocab <- letters
resp <- captcha_transform_label(c("a","b","c","d","e"), vocab)
class(resp)
dim(resp)

}
