% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decrypt.R
\name{decrypt}
\alias{decrypt}
\title{Function to solve Captchas}
\usage{
decrypt(files, model)
}
\arguments{
\item{files}{files to read. Can be a character vector or an object of class \code{captcha}.}

\item{model}{model of class \code{luz_module_fitted}}
}
\description{
Function to solve Captchas
}
