% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.R
\name{new_captcha}
\alias{new_captcha}
\title{Create a new project to solve a new Captcha}
\usage{
new_captcha(path, rstudio = TRUE, open = rstudio)
}
\arguments{
\item{path}{A path. If it exists, it is used. If it does not exist,
it is created, provided that the parent path exists.}

\item{rstudio}{Use RStudio? Defaults to \code{TRUE}.}

\item{open}{Open new RStudio project? Defaults to \code{rstudio}.}
}
\value{
Single logical value indicating if current session is modified.
}
\description{
This function creates a new project with a template implemented inside
the captcha package. It creates a new folder containing helpers to
download, annotate, fit a custom model and share the results.
}
\details{
After creating a new project, via the new_captcha() command or via the
RStudio interface, a new window opens. The project contains four files:
\itemize{
\item 01_download.R: Contains some code to help writing functions that
download Captchas in a real scenario. In practice, the download
functions need to be adapted because the websites are organized
in very different ways.
\item 02_annotate.R: Contains a template for manual annotation of Captchas.
Manual annotation can either be performed using the interface created
by the captcha package or externally. The annotated files are stored
in the img folder.
\item 03_model.R: Contains a template for modeling, allowing complete
customization of the fitting procedure. The script contains commands
to load the data, specify the model, fit the model and save the
fitted model.
\item 04_share.R: Contains operations to create a git repository of the
solution and make the fitted model available. The model can be loaded
afterwards using the captcha_load_model() function, without the need
to copy files locally.
}

For more details, see \code{vignette("advanced")}.
}
